added instant transmit on ACK update, with currentPdu awareness, and removed receive button (auto receive)

+++ app/src/main/java/test/com/blootoothtester/util/Constants.java
@@ -7,6 +7,6 @@ package test.com.blootoothtester.util;
 public class Constants {
     public static final int PDU_ACK_MOD = 127;
     public static final int MAX_USERS = 40;
-    public static final int PDU_BROADCAST_ADDR = MAX_USERS + 1;
+    public static final byte PDU_BROADCAST_ADDR = MAX_USERS + 1;
     public static final int EPOCH_LENGTH_MILLIS = 16000;
 }
